feat: add firsts endpoint to: POST api/users GET api/users/{utgid}

+++ backend/internal/interfaces/http/dto/action_response.go
@@ -4,7 +4,7 @@ import "github.com/isiyar/daily-energy/backend/internal/domain/models"
 
 type ActionResponse struct {
 	Id           string            `json:"id"`
-	Utgid        int               `json:"utgid"`
+	Utgid        int64             `json:"utgid"`
 	Date         int64             `json:"date"`
 	ActivityName string            `json:"activity_name"`
 	Calories     int               `json:"calories"`
